requirements.txt
Django==3.0.7
django-allauth==0.42.0
django-crispy-forms==1.9.1
Pillow==7.1.1
requests==2.23.0
python-decouple==3.3
